+++ src-tauri/capabilities/default.json
@@ -10,6 +10,7 @@
     "opener:default",
     "fs:default",
     "dialog:default",
-    "log:default"
+    "log:default",
+    "fs:default"
   ]
 }
