+++ src/app.css
@@ -6,11 +6,39 @@
 @import "@skeletonlabs/skeleton/optional/presets";
 @import "@skeletonlabs/skeleton/themes/wintry";
 
+@font-face {
+    font-family: "Lobster";
+    src: url("/fonts/Lobster.ttf") format("truetype");
+    font-weight: normal;
+    font-style: normal;
+    font-display: swap;
+}
+
+@font-face {
+    font-family: "Marker Gothic";
+    src: url("/fonts/MarkerGothic.ttf") format("truetype");
+    font-weight: normal;
+    font-style: normal;
+    font-display: swap;
+}
+
+@font-face {
+    font-family: "Oswald";
+    src: url("/fonts/Oswald.ttf") format("truetype");
+    font-weight: normal;
+    font-style: normal;
+    font-display: swap;
+}
+
 @theme inline {
     --radius-sm: calc(var(--radius) - 4px);
     --radius-md: calc(var(--radius) - 2px);
     --radius-lg: var(--radius);
     --radius-xl: calc(var(--radius) + 4px);
+
+    --font-lobster: "Lobster";
+    --font-marker-gothic: "Marker Gothic";
+    --font-oswald: "Oswald";
 }
 
 @layer base {
